pricing service refactoring
