+++ prompt/index.js
